+++ src/training/common/trainingsampleset.cpp
@@ -594,7 +594,7 @@ void TrainingSampleSet::SetupFontIdMap() {
   // Number of samples for each font_id.
   std::vector<int> font_counts;
   for (auto &sample : samples_) {
-    const int font_id = sample->font_id();
+    const size_t font_id = sample->font_id();
     while (font_id >= font_counts.size()) {
       font_counts.push_back(0);
     }
